+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # master
 
 * [DEPRECATION] use `fulfilled` instead of `resolved` [#9](https://github.com/dockyard/ember-cli-async-button/pull/9)
+* Updated addon for ember-cli 0.0.41 [#10](https://github.com/dockyard/ember-cli-async-button/pull/10)